+++ src/wrapper/level_progressions.rs
@@ -12,7 +12,7 @@ use wanikani_rs_model::level_progression::LevelProgression;
 #[derive(Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
 pub enum LevelProgressionFilter {
-    Ids(Vec<i32>),
+    Ids(Vec<i64>),
     UpdatedAfter(DateTime<Utc>),
 }
 
@@ -32,6 +32,6 @@ impl WanikaniClient {
         get_level_progression,
         "level_progressions/{id}",
         ResourceResponse<LevelProgression>,
-        id: i32
+        id: i64
     );
 }
